requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-community>=0.0.10
litellm>=1.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
rich>=13.0.0